% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{vg_fun_kPa}
\alias{vg_fun_kPa}
\title{Van Genuchten water retention function (kPa)}
\usage{
vg_fun_kPa(h_kPa, theta_r, theta_s, alpha, n)
}
\arguments{
\item{h_kPa}{Numeric vector of suction (kPa).}

\item{theta_r}{Residual water content (m\eqn{^3}/m\eqn{^3}).}

\item{theta_s}{Saturated water content (porosity) (m\eqn{^3}/m\eqn{^3}).}

\item{alpha}{VG parameter \eqn{\alpha} in kPa\eqn{^{-1}}.}

\item{n}{VG parameter \eqn{n} (dimensionless, \eqn{> 1}).}
}
\value{
Numeric vector \eqn{\theta(h)}.
}
\description{
Computes volumetric water content \eqn{\theta(h)} for suction \eqn{h} in \strong{kPa}
using the Van Genuchten model with the Mualem linkage \eqn{m = 1 - 1/n}.
}
\examples{
vg_fun_kPa(10, theta_r=0.05, theta_s=0.45, alpha=0.05, n=1.6)
}
