% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{vg_sse_kPa}
\alias{vg_sse_kPa}
\title{VG least-squares objective (internal)}
\usage{
vg_sse_kPa(par, h_kPa, theta_obs)
}
\arguments{
\item{par}{Numeric vector: \code{c(theta_r, theta_s, alpha, n)}.}

\item{h_kPa}{Suction in kPa.}

\item{theta_obs}{Observed \eqn{\theta}.}
}
\description{
Sum of squared errors of \eqn{\theta} for VG parameters.
}
\keyword{internal}
